Perf round 3: unsigned tail folds and incremental candidate refill
Kernel 0.822 s -> 0.719 s (12.5%) on the reference workload; wall
~0.94 s, 205x the as-shipped Java reference, 4.9x the original CUDA
version. Modern generator tail rewritten in provably-unsigned form
(value-identical; compiler folds the constant mod/shifts instead of
emitting signed fixups); refill path uses a per-chunk pull budget and
incremental linear index/x-hash-term so per-candidate 64-bit math and
multiplies survive only on the rare z-carry path. Re-validated: oracle
diff 0/29,178,112, all 29 e2e checks on the GPU, registers unchanged.
Benchmarked on the replacement instance after confirming the previous
build reproduces its numbers there (0.821-0.824 s vs 0.822-0.826 s).

+++ include/texture.cuh
@@ -74,8 +74,12 @@ TF_HOST_DEVICE inline int modernFromCoordRandom(int64_t coordRand, int mod)
     constexpr int64_t MASK = (1LL << 48) - 1;
     int64_t seed = coordRand ^ MULTIPLIER;
     seed = (int64_t)((uint64_t)seed * (uint64_t)MULTIPLIER + 11ull) & MASK;
-    int next = (int)(seed >> (48 - 31));
-    return (int)(((4LL * next) >> 31) % mod);
+    // seed is in [0, 2^48) after the mask, so `next` is non-negative and the
+    // Java tail is computed in unsigned form (identical values for all
+    // inputs; this lets the compiler fold the shift and constant mod instead
+    // of emitting signed-division fixups it can't prove away).
+    unsigned int next = (unsigned int)((uint64_t)seed >> (48 - 31));
+    return (int)(((4ull * next) >> 31) % (unsigned int)mod);
 }
 
 TF_HOST_DEVICE inline int getTextureModern(int32_t x, int32_t y, int32_t z, int mod)
